refactor: move withApiClientAndPagination into own file
