+++ src/pages/admin/locations/[identifier].tsx
@@ -1,6 +1,6 @@
 import { Location } from "@api/client/models/Location";
 import LocationDetailsPage from "@components/pages/LocationDetailsPage";
-import { withApiClientAndPagination } from "@services/data";
+import withApiClientAndPagination from "@services/withApiClientAndPagination";
 import withAuth from "@services/withAuth";
 import { GetServerSideProps } from "next";
