+++ src/pages/admin/organizations/[identifier].tsx
@@ -1,7 +1,7 @@
 import { Organization } from "@api/client/models/Organization";
 import { Tag } from "@common/types";
 import OrganizationDetailsPage from "@components/pages/OrganizationDetailsPage";
-import { withApiClientAndPagination } from "@services/data";
+import withApiClientAndPagination from "@services/withApiClientAndPagination";
 import withAuth from "@services/withAuth";
 import { GetServerSideProps } from "next";
