+++ src/pages/admin/organizations/create.tsx
@@ -1,6 +1,6 @@
 import OrganizationDetailsPage from "@components/pages/OrganizationDetailsPage";
-import { withApiClientAndPagination } from "@services/data";
 import { loadMessages } from "@services/i18n";
+import withApiClientAndPagination from "@services/withApiClientAndPagination";
 import withAuth from "@services/withAuth";
 import { GetServerSideProps } from "next";
