+++ src/pages/admin/users/index.tsx
@@ -1,7 +1,7 @@
 import { Membership } from "@common/types";
 import UsersPage from "@components/pages/UsersPage";
 import { decodeAccessToken } from "@services/auth";
-import { withApiClientAndPagination } from "@services/data";
+import withApiClientAndPagination from "@services/withApiClientAndPagination";
 import withAuth from "@services/withAuth";
 import { GetServerSideProps } from "next";
