+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { Attraction } from "@api/client/models/Attraction";
 import AttractionsPage from "@components/pages/AttractionsPage";
 import { PaginationType } from "@components/Pagination";
-import { withApiClientAndPagination } from "@services/data";
+import withApiClientAndPagination from "@services/withApiClientAndPagination";
 import withAuth from "@services/withAuth";
 import { GetServerSideProps } from "next";
